requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
marshmallow==3.19.0
marshmallow-sqlalchemy==0.28.0
PyMySQL==1.0.3
Werkzeug==2.2.3
python-dotenv==1.0.0
Flask-Migrate==4.0.4
passlib==1.7.4
bcrypt==4.0.1
PyJWT==2.4.0
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4